+++ doc/changes.rst
@@ -7,6 +7,8 @@ desispec Change Log
 
 * Update (non-essential) transfer script for spectrograph functional
   verification tests (PR `#758`_).
+* New calibration data access (inc var. DESI_SPECTRO_CALIB
+  replacing DESI_CCD_CALIBRATION_DATA).
 
 .. _`#758`: https://github.com/desihub/desispec/pull/758